requirements.txt
aiofiles>=23.2.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
beautifulsoup4>=4.12.0
fastapi>=0.104.1
fitz
langchain>=0.1.0
langchain-openai>=0.0.5
lxml
matplotlib>=3.7.0
networkx>=3.1
numpy>=1.24.0
openai>=1.3.0
pandas>=2.0.0
passlib[bcrypt]>=1.7.4
pdfplumber
Pillow
pydantic>=2.5.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
regex
requests>=2.31.0
scikit-learn>=1.3.0
seaborn>=0.12.0
spacy>=3.7.0
tqdm>=4.66.0
trafilatura
urllib3
uvicorn[standard]>=0.24.0
pytest>=8.0.0
pytest-cov>=5.0.0
